:recycle: Tests for http package
